Minor improvement to draw-hatch action
Hatch action failed to properly hatch closed line-boundaries within a closed line-boundary (or extending into a stub or another closed boundary). This was identified as bug and originates in RS_EntityContainer::optimizeContours(), which adds the last line to the contour to be hatched even if its vpEnd=eTmp->getEndpoint(); distorts the hatch shape. The fix is, in such cases, a made-up line linking the lose end back to the point of closure.
Also, RS_Hatch::validate()'s omission of setting needOptimization caused calling optimizeContours() again unecessarily.